Refactor: MyPageController
