+++ Effi-BE-fin/src/main/java/com/example/effi/controller/MyPageController.java
@@ -3,7 +3,6 @@ package com.example.effi.controller;
 
 import com.example.effi.domain.DTO.MyPageResponseDTO;
 import com.example.effi.service.MyPageService;
-import com.example.effi.service.MyPageServiceImpl;
 import lombok.AllArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -17,7 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/mypage")
 public class MyPageController {
 
-    private final MyPageServiceImpl myPageService;
+    private final MyPageService myPageService;
 
     @GetMapping(value = "/{empNo}", produces = {MediaType.APPLICATION_JSON_VALUE})
     public ResponseEntity<MyPageResponseDTO> mypageView(@PathVariable("empNo") Long empNo){
