+++ Effi-BE-fin/src/main/resources/application.properties
@@ -6,7 +6,7 @@ spring.datasource.url=jdbc:mariadb://localhost:3306/effi?createDatabaseIfNotExis
 
 spring.datasource.username=root
 
-spring.datasource.password=2585
+spring.datasource.password=123456789
 
 spring.jpa.hibernate.ddl-auto=update
